+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+### 6.2.3
+
+- Updates `getFeaturesByCoordinate` in `wmsFeatureProvider` to consider the extent
+- adds `htmlPositionFeatureTitle` to `wmsFeatureProvider` which is added as to the point feature created if (responseType === 'text/html').
+
 ### 6.2.2
 
 - Fixes an issue where the first panorama image was not shown in the correct resolution